modularize complex ops
