+++ tfjs-core/src/ops/ops.ts
@@ -27,6 +27,7 @@ export {batchNorm3d} from './batchnorm3d';
 export {batchNorm4d} from './batchnorm4d';
 export {broadcastTo} from './broadcast_to';
 export {clone} from './clone';
+export {complex} from './complex';
 export {concat} from './concat';
 export {concat1d} from './concat_1d';
 export {concat2d} from './concat_2d';
@@ -48,6 +49,7 @@ export {equal} from './equal';
 export {eye} from './eye';
 export {greater} from './greater';
 export {greaterEqual} from './greater_equal';
+export {imag} from './imag';
 export {less} from './less';
 export {lessEqual} from './less_equal';
 export {localResponseNormalization} from './local_response_normalization';
@@ -71,6 +73,7 @@ export {rand} from './rand';
 export {randomGamma} from './random_gamma';
 export {randomNormal} from './random_normal';
 export {randomUniform} from './random_uniform';
+export {real} from './real';
 export {relu} from './relu';
 export {separableConv2d} from './separable_conv2d';
 export {spaceToBatchND} from './space_to_batch_nd';
@@ -82,7 +85,6 @@ export {tile} from './tile';
 export {truncatedNormal} from './truncated_normal';
 
 export * from './boolean_mask';
-export * from './complex_ops';
 export * from './reverse';
 export * from './slice';
 export * from './unary_ops';
